requirements.txt
flask
flask-mysqldb
flask-mail
werkzeug
mysqlclient
python-dotenv